requirements.txt
scikit-learn
pandas
joblib
python-docx
pdfminer.six